Add role manager interface and class
